sbin: ifupdown2d: daemon code
This file takes care of the daemonizing part then will listen, on a UNIX domain
socket, for client requests.
For each request we will, for now, execute the whole "ifupdown2" path/flow. We
currently execute ifupdown2 in a fork to avoid conflict and have a stable code
it will be removed in future fixes.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>