ifupdown2d: refactoring imports to prevent nlmanager instanciation before daemon state
I was stuck on a very strange issue a day or two. Long story short, when I
enable ifupdown2 to become a daemon, the socket returned by socket.accept()
has the same FD has the netlink socket opened by nlmanager... I was printing
the value of both socket, next to each other, I could see the same FDs... which
is... not supposed to happen... On my VM the client was receiving garbage but
when testing on my switch nlmanager would fail with "addr already in use"..
so so weird....

To fix this bug I changed the imports, so that nlmanager doesn't get
instanciated before ifupdown2 becomes a daemon. This bug was probably due to
side effects related to daemon/fork and the use of socket.FD_CLOECEC

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>