Revert "netlink: moving ip link set master commands to netlink"
This reverts commit 3e960ac241768360ddba9751e5472bfc4729e9d3.

Reverting this patch because it causes side effects (see description bellow):

Roopa:
The change itself is ok..., but looking at it closely there are two problems
with this change in the bridge addon module (have not looked at the other addon
modules. Hoping you will help look)

a) The change to netlink was done inside an iproute2 batch command which can
change the order of commands

b) The batch_commit had a try/catch to catch and pretty print the error...which
we missed. And the pretty-print was done because the previous error from
iproute2 (and now nlmanager) is not intuitive. The move to pretty-print was
done as a bug fix for a complaint from gss. we have exposed the non-intuitive
error msg again.

Julien:
Maybe we should completely revert the patch for now and keep it off until we
have all iproute2 batch commands moved to netlink. Otherwise, we will have a
mix of iproute2 and netlink command which can me confusing.