python-nlmanager: Lower syslog level of "Add support for encoding/decoding.." messages
Ticket: CM-16650
Reviewed By: CCR-6384
Testing Done: Ran smoke

nlpacket includes some debug level log messages which warn when an attribute
cannot be encoded or decoded. These message overwhelm the log file and can occur
at such a rate that other log messages are dropped. The log file also rotates
to the point that other messages are rotated off into the ether.

This commit lowers the syslog level of the "Add support for..." messages to be
5, which is lower than debug (10). Therefore, when enabling debug the other
debug log messages can be seen. If one desires to see these "Add support for..."
messages, the log level can be set to 5 or less.

Signed-off-by: Scott Emery <scotte@cumulusnetworks.com>
Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>