addons: address: add support for scope link addresses
Ticket: CM-16290
Reviewed By: Roopa, Nikhil G, Scott Laffer
Testing Done: ifupdown2-tests, smoke-nightly ifupdown2, smoke-nightly ifreload

We now support addr with link scope. Since the kernel may add it's
own link address to some interfaces we need to filter them out and
make sure we only deal with the addresses set by ifupdown2.

To do so we look at the previous configuration made by ifupdown2
(with the help of the statemanager) together with the addresses
specified by the user in /etc/network/interfaces, these addresses
are then compared to the running state of the intf (ip addr show)
made via a netlink addr dump.
For each configured addresses of scope link, we check if it was
previously configured by ifupdown2 to create a final set of the
addresses watched by ifupdown2

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>