addons: address: fix ip6 check, if a running address is not in the user config
Ticket: CM-18599
Reviewed By: Roopa, Nikhil G
Testing Done: ifupdown2-tests, smoke/nightly

This commit fixes a small mistake, this check return True most of the time
if we had multiple ip6. It was checking the user config against the running
ip6 addresses, but the right way is to checj the running addrs against the
user config.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>