addons: bridge: mcsnoop: reset to default when removed from config
Ticket: CM-11389
Reviewed By: Roopa, Nikhil G
Testing Done: ifupdown2-tests, ifupdown2/ifreload smoke/nightly, l2 smoke

$ ifquery br0
auto br0
iface br0
      bridge-mcsnoop 0
      bridge-ports swp1 swp2

$ ifup br0
$ cat /sys/class/net/br0/bridge/multicast_snooping
0
$ ifquery br0
auto br0
iface br0
      bridge-ports swp1 swp2

$ ifreload -a
$ cat /sys/class/net/br0/bridge/multicast_snooping
1
$

bridge-mcsnoop was processed twice (once in the main loop because it was part
of the map '_ifla_br_attributes_map' and once after that for special handling
I commented out the map entry to avoid the double processing.
Also we werent resetting this attribute because of the 'if' check before
processing mcsnoop and also because the cached value was 'False' and was considered as Null and thus not entering the - back to default - path.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>