LinkUtils:: only refresh bond and bridge attr for the provided ifname
Ticket: 0
Reviewed By: Roopa, Nikhil G
Testing Done: ifupdown2-tests

Today if we want to refresh the cache, all the bond and bridge info
will be refreshed, even if a specific ifname was provided. This patch
fixes this behavior by only refreshing the requested interface.
This should improve perf a bit (and comply with the old behavior)

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>