ifupdownaddons: iproute2: fix 'scope' address exception
Ticket: CM-18647, CM-19279
Reviewed By: Roopa, Nikhil G
Testing Done: ifupdown2-tests

When macvlans are created, we manually add their addr to the cache.
The cache usually relies on a kernel dump with all the info (including
scope), in that case scope is missing.
Adding a try/except to do specific handling in that case.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>