LinkUtils: get_running_addrs: fix variable re-declaration 'details'
Ticket: 0
Reviewed By: Roopa, Nikhil G
Testing Done:

today I hit a case were I got an unexpected KeyError exception 0 in the scheduler
I traced it back to the get_running_addrs function returning a dictionary instead
of a list, even when the last parameter - details=False - was provided. With this
parameter set to false we should get a list and not a dict.

I realized that when entering the special condition (address scope not present in
the cache) the 'details' variable would be reset to {} then filled.

This is a uncaught regression.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>