networking.service wasn't running at shutdown
Ticket: CM-19638
Reviewed By: roopa,wkok,julien
Testing Done: Added logger commands to script, also watched console

In the dependency cleanups in 3.5 when working with the new systemd,
I broke some things when I added DefaultDependencies=no to some
services.

mstpd and cumulus-fastfailover are also affected by this, so I'm
fixing it under the same bug.

The fix is to add Before=shutdown.target and Conflicts=shutdown.target