addons: vlan: warn if vlan-protocol attr is removed from stanza
Ticket: CM-18144
Reviewed By: Roopa, Nikhil G, Anitha
Testing Done: test explained in the bug

[6:20:44] root:~ # ifquery swp3.11
auto swp3.11
iface swp3.11
      vlan-protocol 802.1ad

[6:20:50] root:~ # ifreload -a
[6:21:03] root:~ #
[6:21:08] root:~ # ifquery swp3.11
auto swp3.11
iface swp3.11
      vlan-protocol 802.1q

[6:21:10] root:~ # ifreload -a
error: swp3.11: cannot change vlan-protocol to 802.1q: operation not supported. Please delete the device with 'ifdown swp3.11' and recreate it to apply the change.
[6:21:11] root:~ #
[6:21:13] root:~ #
[6:21:13] root:~ # ifdown -a -X eth0
[6:21:14] root:~ # ifquery swp3.11
auto swp3.11
iface swp3.11
      vlan-protocol 802.1q

[6:21:17] root:~ # ifreload -a
[6:21:19] root:~ # ifquery swp3.11
auto swp3.11
iface swp3.11
      vlan-protocol 802.1ad

[6:21:24] root:~ # ifreload -a
error: swp3.11: cannot change vlan-protocol to 802.1ad: operation not supported. Please delete the device with 'ifdown swp3.11' and recreate it to apply the change.
[6:21:25] root:~ # echo $?
1
[6:21:27] root:~ #

When removing an attribute from a stanza the rule is to reset to default.
vlan-protocol can't be change therefore we need to warn the user if the
vlan-protocol line is removed and the running protocol is different from the
default defined in modinfo.

Ideally if vlan-protocol is specified in the policy files we should the
policy value as default. To be discussed...

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>