addons: address: don't mix ip objects with strings
newaddrs was filled with utils.get_ip_objs and is passed to
self._inet_address_list_config() in two places. In one of
them it conditionally gets an anycast address as string
before being converted to a list of strings via
self.order_user_configured_addrs() (which tests for IPv6
addresses via `isinstance()`).

While the anycast address is `clagd-vxlan-anycast-ip` which
is supposedly only ipv4, the correct ordering is more of a
side effect than an obvious intention, and we want to be
able to use the `isinstance()` in
self._inet_address_list_config() later on again, so strictly
separate the two different lists.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>