addons: address: explicitly set 'disable_ipv6'
When the user requests an interface to get an IPv6 address
we need to make sure net.ipv6.conf.$iface.disable_ipv6=0 in
order for the address assignment to work. Othrewise we
cannot bring up bridges with ipv6 if
net.ipv6.conf.default.disable_ipv6=1.

When no IPv6 address is requested, we reset the value to
net.ipv6.conf.default.disable_ipv6.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>