lib: new base_objects.py io.py and sysfs.py
Beginning of ifupdown2 cleanup: the project needs a rework of most of the
base classes. Today we have a big mess of ifupdownaddons, utils (mstpctlutils,
systemutils) and many different base classes (modulebase, utilsbase,
ifupdownbase etc..). All of it needs to be refactored, and ported to the new
netlink cache. DRYRUN needs to be handled by the the DryRun manager.

base_objects.py: contains BaseObject, most of ifupdown2 objects should
subclass BaseObject as it will handle dry-run handlers and dry-run context.
As well as instantiating a named logger (self.logger).
It also contains Cache and Netlink that directly subclass BaseObject
and adds respectively an access to the global netlink cache (self.cache)
as well as a reference to the global netlink API (self.netlink).

Since some of the modules don't all require access to both cache and API
it was decided to separate them so users can easily create a custom
subclass and subclass any (or multiple) class from base_objects.py

io.py: aims at centralizing all IO related functions

sysfs.py: subclass IO and Cache so it can manipulate the kernel filesystem
but also access the cache at the same time.

that's all for now but other base classes will be introduced:
- OS operation (subprocess execution, PID checks etc)
and probably more...

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>