addons: address: move mtu handling to sysfs
Today MTU is handled by a mix of oldcache with sysfs and iproute2. When
taking a closer look at scale config we can see that ifupdown2 spends a
lot of time executing iproute2 cmd: ip link set dev DEV mtu MTU takes up
to several dozens of ms to run.
The goal of this patch is to move all the mtu code to use the new netlink
cache as well as using sysfs sets to set devices MTU. sysfs is indeed
the best way to set the MTU today (running at under 1ms) while a netlink
set request would take somewhere between 10 and 20ms. This is the most
efficient way to set MTU today. We won't be able to do better until
we have a way to batch netlink packets.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>