lib: nlcache: add get_link_attribute_raw and get_link_address_raw API
Those API are used to get the raw data of an attribute for example:
IFLA_ADDRESS (mac address), in this case the mac is represented by an
integer. We both store this value as integer and string. ifupdown2 needs
to be able to query the cache and get the integer value for comparaisons

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>