lib: nlcache: address cache now splits ip4 and ip6 in separate sub data structures
ifupdown2 needs to store ip addresses in family specific lists.
The new address_cache has the following schema:

_addr_cache = {
    swp42: {
        4: [nlpacket.Address(IPv4), ],
	6: [nlpacket.Address(IPv6), ]
    }
}

old schema was:
{
    swp42: [nlpacket.Address, nlpacket.Address, ]
}

When handling the configuration below, ifupdown2 faced several challenges:
auto swp1
iface swp1
      address 2001:db8::1/128
      address 192.0.2.1/32
      address 2001:db8::42/128
      address 192.0.2.42/32
      address 2042:db8::1/128
      address 192.0.42.1/32
      address 192.42.2.42/32

- First, ifupdown2 expects to receive ordered dumps from the kernel, so we
simply populated our list of addresses for each interfaces. For every new
address notification we would append the object to the end of the list.

This config breaks the assumption of having ordered addresses in cache.
ip addr add 2001:db8::1/128 -> add the ip6 first in the list
ip addr add 192.0.2.1/32 -> the primary ip4 is then second...

Fix: move to a family specific data structure.

- Second, it seems like when ip6 are added we get two notifications, one
for the creation and a second notification for the same address but with
updated flags and cacheinfo..

Fix: check if the address is already cached, if so, simple update its cache
entry with the new notification.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>