daemon: better handle client-daemon communication via socket handler
A serious issues apparead when running ifupdown2d with debug output
enabled. After processing the client request we restore the original
daemon log-level (before emitting the exit status of the request).
Then we close the socket handler and remove it from the root logger.

For some unknown (yet) reasons if the daemon emit log messages (from
a thread) while we try to remove and close the socket handler, the
client never receive any connection from the daemon (thus blocking
and hanging indefinitly).

To solve this, first we shouldn't restore the daemon log-level, for
now we will simply set it to INFO so that we can emit the request
exit status to syslog. Then we remove the socket handler from the
root logger before doing anything else, that way message logged on
the root logger shouldn't be dispatched to the socket handler. Now
we can close it's socket and make sure we reset the retryTime.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>