nlcache: add support for bridge vlans - replace all LinkUtils apis
The netlink cache now supports bridge vlans (AF_BRIDGE packets). On RX
RTM_NEWLINK we process AF_BRIDGE packets to extract vlans and flags from
IFLA_BRIDGE_VLAN_INFO (if provided). Those vlans are then stored in a
new data structure: _bridge_vlan_cache.

To properly implement this, some existing code had to be refactored. The
remove/del actions on the masters_slaves dictionary are now encapsulated
in the "__unslave" function. This is necessary to make sure that when
unslaving a device the associated bridge_vlan entry is also cleared.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>