addons: vxlan: refactoring up ops to avoid resending all attributes on changes
The vxlan module is due for a big refactoring. The current code is
very messy and hard to get around. Now each attribute has its own
subfunction to get the user config and/or policy value then translate
it to a netlink value saved in a dictionary that we inject in our
VXLAN add/set request.

Ticket: CM-24735
Reviewed-by: CCR-8593

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>