lib: dry_run: fix DRY-RUN handler using weakrefs
We disabled DRY-RUN mode because it was introducing huge memory leaks. The
cause of this issue was that we kept strong references on objects/modules
and methods in our internal data-structure.

To solve this problem we need to move all the existing dry-run handling code
to use weakrefs. Then another problem emerged ; you can't use the weakref
module on bound methods. To solve this last issue we had to get some inspira
tion from ActiveState recipe 81253, implementing customs class to hold
weakrefs on bound methods but using a ingenious trick to call the dry-run
method when needed.

Signed-off-by: Julien Fortin <julien@cumulusnetworks.com>