address: add cache.get_link_inet6_accept_ra and autoconf
factorize with adding_get_netlink_cache_accept_ra && _get_netlink_cache_auto
and correctly test if cache exist or not

the netlink cache can have empty value for an interface,
if the interface was not existing when we have populate the cache.
for example, a vlan interface created in pre-up by vlan module.

In this case, we return an empty string
, and fix parsing of empty netlink cache