Implement new interfaces selection for ifreload
Reload operation uses a -a/--all option without --allow auto as default.
This command has an impact on ALL interfaces by default.

The --allow can be set to filter the targetted interfaces.

I do not really understand why the IFLIST is disable on this operation.
It could be very useful.