Parsing integration for all ifupdown2 operations
Use the new methods to create a standard interface selection for
ifup, ifdown, ifquery and ifreload operations.

ifquery and ifreload have a tricky implicit configuration, but I have
tried to not modify the ifupdown2 behavior in any way.

There is one changement on the -a/--all option that can be present or not
when implicit. Now, the `ifreload -a -c` command works for example.